# openapi/planning-activities.yaml
openapi: 3.1.0
info:
  title: Planning Activities API
  version: 0.1.0
  description: |
    REST-Schnittstelle für die Gantt-Aktivitäten der Order-Management-Anwendung.
    Alle Clients arbeiten gegen denselben Datenbestand. Validierungen für Orts-,
    Kapazitäts-, Arbeitszeit- und Qualifikationskonflikte werden serverseitig
    ausgeführt und können gezielt angefordert werden.
servers:
  - url: /api/v1
    description: Default Backend-URL (überschreibbar per API_CONFIG).
paths:
...
  /api/templates:
    get:
      summary: Templates auflisten
      description: Liefert Metadaten aller Templates (eigene Timeline-Tabellen).
      responses:
        '200':
          description: Liste der Templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: './planning/components/schemas.yaml#/schemas/TemplateSet'
    post:
      summary: Template anlegen
      description: Legt ein neues Template an und erzeugt die zugehörige Timeline-Tabelle.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './planning/components/schemas.yaml#/schemas/CreateTemplateSetRequest'
      responses:
        '201':
          description: Angelegtes Template
          content:
            application/json:
              schema:
                $ref: './planning/components/schemas.yaml#/schemas/TemplateSet'
  /api/templates/{templateId}:
    get:
      summary: Template laden
      parameters:
        - in: path
          name: templateId
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Template-Metadaten
          content:
            application/json:
              schema:
                $ref: './planning/components/schemas.yaml#/schemas/TemplateSet'
    put:
      summary: Template aktualisieren
      parameters:
        - in: path
          name: templateId
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: './planning/components/schemas.yaml#/schemas/UpdateTemplateSetRequest'
      responses:
        '200':
          description: Aktualisiertes Template
          content:
            application/json:
              schema:
                $ref: './planning/components/schemas.yaml#/schemas/TemplateSet'
    delete:
      summary: Template löschen
      parameters:
        - in: path
          name: templateId
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Gelöscht, zugehörige Timeline-Tabelle entfernt
...